+++ src/elite/libs/EliteSharpLib/BreakPattern.cs
@@ -1,4 +1,4 @@
-// 'Elite - The Sharp Kind' - Andy Hawkins 2023-2026.
+﻿// 'Elite - The Sharp Kind' - Andy Hawkins 2023-2026.
 // 'Elite - The New Kind' - C.J.Pinder 1999-2001.
 // Elite (C) I.Bell & D.Braben 1984.
 
@@ -47,7 +47,6 @@ internal sealed class BreakPattern
 
     internal void Reset()
     {
-        _draw.SetViewClipRegion();
         _breakPatternCount = 0;
         IsComplete = false;
     }